bug with update duration for memory pool objects fixed. user session's current user now uses the memory pool. set object method of memory pool refreshes the updated at field for the pool object being set
git-svn-id: svn+ssh://svn.denwen.com/vol/svn/mobile/iphone/branches/experimental@658 ab2db6a2-7c84-486d-9f3d-ac5f155663ef

+++ Denwen/Classes/DWConstants.h
@@ -274,7 +274,7 @@ extern NSInteger const kMPPlacesIndex;
 extern NSInteger const kMPUsersIndex;
 extern NSInteger const kMPAttachmentsIndex;
 extern NSInteger const kMPAttachmentSlicesIndex;
-extern NSInteger const kMPObjectUpdateInterval;
+extern float     const kMPObjectUpdateInterval;
 extern NSInteger const kMPDefaultDatabaseID;
